+++ app/src/main/java/com/juergenkleck/android/game/hedgehog/engine/HedgehogEngine.java
@@ -1,0 +1,12 @@
+package com.juergenkleck.android.game.hedgehog.engine;
+
+import com.juergenkleck.android.gameengine.engine.GameEngine;
+
+/**
+ * Android app - Hedgehog
+ *
+ * Copyright 2022 by Juergen Kleck <develop@juergenkleck.com>
+ */
+public interface HedgehogEngine extends GameEngine {
+
+}
