+++ app/src/main/java/com/juergenkleck/android/game/hedgehog/rendering/objects/Obstacle.java
@@ -1,0 +1,21 @@
+package com.juergenkleck.android.game.hedgehog.rendering.objects;
+
+/**
+ * Android app - Hedgehog
+ *
+ * Copyright 2022 by Juergen Kleck <develop@juergenkleck.com>
+ */
+public class Obstacle extends com.juergenkleck.android.gameengine.rendering.objects.Obstacle {
+
+    public Obstacle(int type, float life, int gReference, int points) {
+        super(type, life, gReference, points);
+    }
+
+    // movement time
+    public long lastMove;
+
+    // the obstacle position on the game map
+    public int x;
+    public int y;
+
+}
