--- app/src/main/java/info/simplyapps/game/hedgehog/engine/HedgehogEngine.java
@@ -1,7 +1,0 @@
-package info.simplyapps.game.hedgehog.engine;
-
-import info.simplyapps.gameengine.engine.GameEngine;
-
-public interface HedgehogEngine extends GameEngine {
-
-}
